.js - random picker now won't open duplicates
It is now impossible, in one page session, to get the same grid again. This is for the 1 in ~30k chance that the randomiser picks the same number in row and so the user would get the same puzzle.

Also increased max grids from 30,000 to 30,380.
Also commented a lot of code.